support for structured json in lang files
Adjust the lang file parsing to allow nesting of keys. Generate
translation keys as paths based on the structure of the json. For
example, given this JSON:
  {
    "foo": {
      "bar": "baz"
    },
    "yay": "wohoo!"
  }

This can *kind of* already be achieved with the current parser as:
  {
    "foo/bar": "baz",
    "yay": "wohoo!",
  }

However, this lacks structure and is harder to read, especially if
desired structure has multiple levels of nesting. The structured version
has the added benefit of supporting code folding in editors, helping
when working with larger sets of translation keys.

The example JSONs for the existing parser and for the updated parser
both generate identical translation keys:
  domain:foo/bar => "baz"
  domain:yay => "wohoo!"

Therefore, the changes are mostly opt-in, giving more flexibility, while
still allowing using the old "flat" lang files as-is.